+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # 1.0.0-beta.6
 
+* **Breaking change:** Chocolatey now uses the `choco` CLI to build and deploy
+  packages, rather than re-implementing its logic in Dart. In particular:
+
+  * The `pkg-chocolatey-build` task has been renamed to `pkg-chocolatey-pack` to
+    match the CLI's naming scheme.
+
+  * The `pkg-chocolatey-pack` and `pkg-chocolatey-deploy` tasks must be run in
+    an environment with the `choco` command available.
+
+* Add a `pkg-chocolatey` command that builds an un-archived Chocolatey package
+  directory.
+
 * Rather than releasing binary snapshots on Chocolatey, compile the source code
   to compile native executables on users' machines.